+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 * Item, Equipment : Added `is_loot`
 * Callback : Fixed `remove` not actually doing anything.
     * Also fixed various `remove_callbacks` that may or may not have been broken.
-* Actor : `fire_` methods now have an optional `no_proc` bool argument (default `false`).
+* Actor : `fire_` methods now have an optional `can_proc` bool argument (default `true`).
     * Also returns the attack instance only now.
 * Helper : Added `is_true` and `is_false`
 * Instance : Fixed `get_data` pulling a different table in certain circumstances.